Change settings entry to use inheritance instead of an enum
